redis_worker_service/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
asyncpg==0.30.0
attrs==25.3.0
backoff==2.2.1
billiard==4.2.1
celery==5.5.3
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
dataclasses-json==0.6.7
distro==1.9.0
frozenlist==1.7.0
fsspec==2024.12.0
greenlet==3.2.3
grpcio==1.73.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
hyperframe==6.1.0
idna==3.10
jiter==0.10.0
json_repair==0.39.1
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.5.4
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.69
langchain-neo4j==0.4.0
langchain-openai==0.3.28
langchain-text-splitters==0.3.8
langsmith==0.4.7
marshmallow==3.26.1
mem0ai==0.1.114
multidict==6.6.3
mypy_extensions==1.1.0
neo4j==5.28.1
neo4j-graphrag==1.8.0
numpy==2.2.6
openai==1.97.0
orjson==3.11.0
packaging==25.0
portalocker==2.10.1
posthog==6.1.1
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pypdf==5.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
qdrant-client==1.14.3
rank-bm25==0.2.2
redis==4.6.0
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rq==2.4.0
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
types-PyYAML==6.0.12.20250516
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.20.1
zstandard==0.23.0
